+++ protocols/ping/tests/ping.rs
@@ -30,9 +30,10 @@ use libp2p::core::{
 use libp2p::mplex;
 use libp2p::noise;
 use libp2p::ping;
-use libp2p::swarm::{behaviour, Swarm, SwarmEvent};
+use libp2p::swarm::{behaviour, NetworkBehaviour, Swarm, SwarmEvent};
 use libp2p::tcp::{GenTcpConfig, TcpTransport};
 use libp2p::yamux;
+use libp2p::NetworkBehaviour;
 use quickcheck::*;
 use rand::prelude::*;
 use std::{num::NonZeroU8, time::Duration};
@@ -40,15 +41,13 @@ use std::{num::NonZeroU8, time::Duration};
 #[test]
 fn ping_pong() {
     fn prop(count: NonZeroU8, muxer: MuxerChoice) {
-        let cfg = ping::Config::new()
-            .with_keep_alive(true)
-            .with_interval(Duration::from_millis(10));
+        let cfg = ping::Config::new().with_interval(Duration::from_millis(10));
 
         let (peer1_id, trans) = mk_transport(muxer);
-        let mut swarm1 = Swarm::new(trans, ping::Behaviour::new(cfg.clone()), peer1_id.clone());
+        let mut swarm1 = Swarm::new(trans, Behaviour::new(cfg.clone()), peer1_id.clone());
 
         let (peer2_id, trans) = mk_transport(muxer);
-        let mut swarm2 = Swarm::new(trans, ping::Behaviour::new(cfg), peer2_id.clone());
+        let mut swarm2 = Swarm::new(trans, Behaviour::new(cfg), peer2_id.clone());
 
         let (mut tx, mut rx) = mpsc::channel::<Multiaddr>(1);
 
@@ -63,16 +62,19 @@ fn ping_pong() {
             loop {
                 match swarm1.select_next_some().await {
                     SwarmEvent::NewListenAddr { address, .. } => tx.send(address).await.unwrap(),
-                    SwarmEvent::Behaviour(ping::Event {
+                    SwarmEvent::Behaviour(BehaviourEvent::Ping(ping::Event {
                         peer,
                         result: Ok(ping::Success::Ping { rtt }),
-                    }) => {
+                    })) => {
                         count1 -= 1;
                         if count1 == 0 {
                             return (pid1.clone(), peer, rtt);
                         }
                     }
-                    SwarmEvent::Behaviour(ping::Event { result: Err(e), .. }) => {
+                    SwarmEvent::Behaviour(BehaviourEvent::Ping(ping::Event {
+                        result: Err(e),
+                        ..
+                    })) => {
                         panic!("Ping failure: {:?}", e)
                     }
                     _ => {}
@@ -86,16 +88,19 @@ fn ping_pong() {
 
             loop {
                 match swarm2.select_next_some().await {
-                    SwarmEvent::Behaviour(ping::Event {
+                    SwarmEvent::Behaviour(BehaviourEvent::Ping(ping::Event {
                         peer,
                         result: Ok(ping::Success::Ping { rtt }),
-                    }) => {
+                    })) => {
                         count2 -= 1;
                         if count2 == 0 {
                             return (pid2.clone(), peer, rtt);
                         }
                     }
-                    SwarmEvent::Behaviour(ping::Event { result: Err(e), .. }) => {
+                    SwarmEvent::Behaviour(BehaviourEvent::Ping(ping::Event {
+                        result: Err(e),
+                        ..
+                    })) => {
                         panic!("Ping failure: {:?}", e)
                     }
                     _ => {}
@@ -118,16 +123,15 @@ fn ping_pong() {
 fn max_failures() {
     fn prop(max_failures: NonZeroU8, muxer: MuxerChoice) {
         let cfg = ping::Config::new()
-            .with_keep_alive(true)
             .with_interval(Duration::from_millis(10))
             .with_timeout(Duration::from_millis(0))
             .with_max_failures(max_failures.into());
 
         let (peer1_id, trans) = mk_transport(muxer);
-        let mut swarm1 = Swarm::new(trans, ping::Behaviour::new(cfg.clone()), peer1_id.clone());
+        let mut swarm1 = Swarm::new(trans, Behaviour::new(cfg.clone()), peer1_id.clone());
 
         let (peer2_id, trans) = mk_transport(muxer);
-        let mut swarm2 = Swarm::new(trans, ping::Behaviour::new(cfg), peer2_id.clone());
+        let mut swarm2 = Swarm::new(trans, Behaviour::new(cfg), peer2_id.clone());
 
         let (mut tx, mut rx) = mpsc::channel::<Multiaddr>(1);
 
@@ -140,13 +144,16 @@ fn max_failures() {
             loop {
                 match swarm1.select_next_some().await {
                     SwarmEvent::NewListenAddr { address, .. } => tx.send(address).await.unwrap(),
-                    SwarmEvent::Behaviour(ping::Event {
+                    SwarmEvent::Behaviour(BehaviourEvent::Ping(ping::Event {
                         result: Ok(ping::Success::Ping { .. }),
                         ..
-                    }) => {
+                    })) => {
                         count1 = 0; // there may be an occasional success
                     }
-                    SwarmEvent::Behaviour(ping::Event { result: Err(_), .. }) => {
+                    SwarmEvent::Behaviour(BehaviourEvent::Ping(ping::Event {
+                        result: Err(_),
+                        ..
+                    })) => {
                         count1 += 1;
                     }
                     SwarmEvent::ConnectionClosed { .. } => return count1,
@@ -162,13 +169,16 @@ fn max_failures() {
 
             loop {
                 match swarm2.select_next_some().await {
-                    SwarmEvent::Behaviour(ping::Event {
+                    SwarmEvent::Behaviour(BehaviourEvent::Ping(ping::Event {
                         result: Ok(ping::Success::Ping { .. }),
                         ..
-                    }) => {
+                    })) => {
                         count2 = 0; // there may be an occasional success
                     }
-                    SwarmEvent::Behaviour(ping::Event { result: Err(_), .. }) => {
+                    SwarmEvent::Behaviour(BehaviourEvent::Ping(ping::Event {
+                        result: Err(_),
+                        ..
+                    })) => {
                         count2 += 1;
                     }
                     SwarmEvent::ConnectionClosed { .. } => return count2,
@@ -191,11 +201,7 @@ fn unsupported_doesnt_fail() {
     let mut swarm1 = Swarm::new(trans, behaviour::KeepAlive, peer1_id.clone());
 
     let (peer2_id, trans) = mk_transport(MuxerChoice::Mplex);
-    let mut swarm2 = Swarm::new(
-        trans,
-        ping::Behaviour::new(ping::Config::new().with_keep_alive(true)),
-        peer2_id.clone(),
-    );
+    let mut swarm2 = Swarm::new(trans, Behaviour::new(ping::Config::new()), peer2_id.clone());
 
     let (mut tx, mut rx) = mpsc::channel::<Multiaddr>(1);
 
@@ -216,10 +222,10 @@ fn unsupported_doesnt_fail() {
 
         loop {
             match swarm2.select_next_some().await {
-                SwarmEvent::Behaviour(ping::Event {
+                SwarmEvent::Behaviour(BehaviourEvent::Ping(ping::Event {
                     result: Err(ping::Failure::Unsupported),
                     ..
-                }) => {
+                })) => {
                     swarm2.disconnect_peer_id(peer1_id).unwrap();
                 }
                 SwarmEvent::ConnectionClosed { cause: Some(e), .. } => {
@@ -261,5 +267,20 @@ enum MuxerChoice {
 impl Arbitrary for MuxerChoice {
     fn arbitrary<G: Gen>(g: &mut G) -> MuxerChoice {
         *[MuxerChoice::Mplex, MuxerChoice::Yamux].choose(g).unwrap()
+    }
+}
+
+#[derive(NetworkBehaviour)]
+struct Behaviour {
+    keep_alive: behaviour::KeepAlive,
+    ping: ping::Behaviour,
+}
+
+impl Behaviour {
+    fn new(config: ping::Config) -> Self {
+        Self {
+            keep_alive: behaviour::KeepAlive,
+            ping: ping::Behaviour::new(config),
+        }
     }
 }