+++ protocols/rendezvous/examples/discover.rs
@@ -26,7 +26,9 @@ use libp2p::ping::{Ping, PingConfig, PingEvent, PingSuccess};
 use libp2p::swarm::SwarmEvent;
 use libp2p::Swarm;
 use libp2p::{development_transport, rendezvous, Multiaddr};
+use libp2p_swarm::behaviour;
 use std::time::Duration;
+use void::Void;
 
 const NAMESPACE: &str = "rendezvous";
 
@@ -44,11 +46,8 @@ async fn main() {
         development_transport(identity.clone()).await.unwrap(),
         MyBehaviour {
             rendezvous: rendezvous::client::Behaviour::new(identity.clone()),
-            ping: Ping::new(
-                PingConfig::new()
-                    .with_interval(Duration::from_secs(1))
-                    .with_keep_alive(true),
-            ),
+            ping: Ping::new(PingConfig::new().with_interval(Duration::from_secs(1))),
+            keep_alive: behaviour::KeepAlive,
         },
         PeerId::from(identity.public()),
     );
@@ -139,10 +138,17 @@ impl From<PingEvent> for MyEvent {
     }
 }
 
+impl From<Void> for MyEvent {
+    fn from(event: Void) -> Self {
+        void::unreachable(event)
+    }
+}
+
 #[derive(libp2p::NetworkBehaviour)]
 #[behaviour(event_process = false)]
 #[behaviour(out_event = "MyEvent")]
 struct MyBehaviour {
     rendezvous: rendezvous::client::Behaviour,
     ping: Ping,
+    keep_alive: behaviour::KeepAlive,
 }