+++ protocols/rendezvous/examples/register_with_identify.rs
@@ -26,7 +26,9 @@ use libp2p::ping::{Ping, PingConfig, PingEvent, PingSuccess};
 use libp2p::swarm::{Swarm, SwarmEvent};
 use libp2p::{development_transport, rendezvous};
 use libp2p::{Multiaddr, NetworkBehaviour};
+use libp2p_swarm::behaviour;
 use std::time::Duration;
+use void::Void;
 
 #[tokio::main]
 async fn main() {
@@ -47,11 +49,8 @@ async fn main() {
                 identity.public(),
             )),
             rendezvous: rendezvous::client::Behaviour::new(identity.clone()),
-            ping: Ping::new(
-                PingConfig::new()
-                    .with_interval(Duration::from_secs(1))
-                    .with_keep_alive(true),
-            ),
+            ping: Ping::new(PingConfig::new().with_interval(Duration::from_secs(1))),
+            keep_alive: behaviour::KeepAlive,
         },
         PeerId::from(identity.public()),
     );
@@ -138,6 +137,12 @@ impl From<PingEvent> for MyEvent {
     }
 }
 
+impl From<Void> for MyEvent {
+    fn from(event: Void) -> Self {
+        void::unreachable(event)
+    }
+}
+
 #[derive(NetworkBehaviour)]
 #[behaviour(event_process = false)]
 #[behaviour(out_event = "MyEvent")]
@@ -145,4 +150,5 @@ struct MyBehaviour {
     identify: Identify,
     rendezvous: rendezvous::client::Behaviour,
     ping: Ping,
+    keep_alive: behaviour::KeepAlive,
 }