+++ swarm/src/behaviour/keep_alive.rs
@@ -6,6 +6,12 @@ use libp2p_core::PeerId;
 use std::task::{Context, Poll};
 
 /// Implementation of [`NetworkBehaviour`] that doesn't do anything other than keep all connections alive.
+///
+/// This is primarily useful for test code. In can however occasionally be useful for production code too.
+/// The caveat is that open connections consume system resources and should thus be shutdown when
+/// they are not in use. Connections can also fail at any time so really, your application should be
+/// designed to establish them when necessary, making the use of this behaviour likely redundant.
+#[derive(Default)]
 pub struct KeepAlive;
 
 impl NetworkBehaviour for KeepAlive {